New API for the gaussian random walk kernel
